Revert "requirements: internalize `packaging.version`"
Since `packaging` is required by `requirements-parser`, it will be installed anyway.

This reverts commit 0a9763aee5c59c3148fbe0edf29fca6888821453.
